requirements.txt
grpcio>=1.59.0
grpcio-tools>=1.59.0
opencv-python>=4.8.1
numpy>=1.26.0
pytest>=7.4.3
setuptools>=69.0.2
protobuf>=4.21.0
pyftpdlib>=1.5.9
